Cleanup and miscellaneous fixes
