+++ frontend/src/components/styles/navbar.css
@@ -53,7 +53,7 @@ nav {
 }
 
 .button-back.rotated {
-  transform: rotate(180deg);
+  transform: rotate(-180deg);
   transition: 0.5s;
 }
