+++ frontend/src/components/styles/submitProblemNotification.css
@@ -1,5 +1,4 @@
 .notification-wrapper {
-  position: absolute;
   width: 100vw;
   height: 100vh;
   display: flex;
