+++ frontend/src/data/optionsData.js
@@ -1,0 +1,16 @@
+const opts = [
+  {
+    optionName: "Close thread",
+    onClick: () => console.log("Thread closed"),
+  },
+  {
+    optionName: "Settings",
+    onClick: () => console.log("Settings opened"),
+  },
+  {
+    optionName: "Change Volunteer",
+    onClick: () => console.log("Volunteer changed"),
+  },
+];
+
+export default opts;
