+++ frontend/src/index.css
@@ -15,31 +15,26 @@ body {
   height: 100vh;
 }
 
-.blurred {
-  position: fixed;
+.modal-container {
+  position: absolute;
   width: 100vw;
   height: 100vh;
-  z-index: 6;
-  left: 0;
-  top: 0;
+  z-index: 300 !important;
   transition: 1s;
 }
 
 
-.blurred.activated {
-  filter: blur(4px) brightness(90%);
-  transition: 1s;
-  z-index: 6;
-  pointer-events: none;
-  user-select: none;
 
+.modal-container.blurred {
+  backdrop-filter: blur(4px) brightness(90%);
+  transition: 1s;
 }
 
 .popup-wrapper {
   position: fixed;
   width: 100vw;
   height: 100vh;
-  z-index: 6;
+  z-index: 7;
   opacity: 0;
   visibility: hidden;
   transition: opacity 0.3s, visibility 0s linear 0.3s;
